refactor: Move indexer RPC tests that should be run in isolation
