+++ crates/iota-indexer/tests/rpc-tests/coin_api.rs
@@ -13,7 +13,6 @@ use iota_json_rpc_types::{
     IotaObjectResponseQuery, IotaTransactionBlockEffectsAPI, IotaTransactionBlockResponse,
     IotaTransactionBlockResponseOptions, IotaTypeTag, ObjectChange, TransactionBlockBytes,
 };
-use iota_keys::keystore::AccountKeystore;
 use iota_move_build::BuildConfig;
 use iota_types::{
     IOTA_FRAMEWORK_ADDRESS, TypeTag,
@@ -33,7 +32,7 @@ use tokio::sync::{Mutex, OnceCell};
 
 use crate::common::{
     ApiTestSetup, execute_tx_and_wait_for_indexer, indexer_wait_for_object,
-    indexer_wait_for_transaction, start_test_cluster_with_read_write_indexer,
+    indexer_wait_for_transaction,
 };
 
 static COMMON_TESTING_ADDR_AND_CUSTOM_COIN_NAME: OnceCell<(IotaAddress, IotaKeyPair, String)> =
@@ -397,105 +396,6 @@ fn fullnode_get_coin_metadata_with_migrated_coin_manager_coins() {
 
         assert!(result_fullnode.is_some());
         assert_eq!(result_fullnode, result_indexer);
-    });
-}
-
-#[test]
-fn indexer_get_coin_metadata_with_migrated_coin_manager_coins() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-    runtime.block_on(async move {
-        let (cluster, store, client) = &start_test_cluster_with_read_write_indexer(
-            Some("indexer_get_coin_metadata_with_migrated_coin_manager_coins"),
-            None,
-            None,
-        )
-        .await;
-
-        let address = cluster.wallet.active_address().unwrap();
-        let address_kp = cluster
-            .wallet
-            .config()
-            .keystore()
-            .get_key(&address)
-            .unwrap();
-        let (coin_name, immutable_metadata_coin_name) =
-            create_migrated_coin_manager_coins(cluster, client, store, address, address_kp)
-                .await
-                .unwrap();
-
-        let (_, result_indexer) =
-            get_coin_metadata_fullnode_indexer(cluster, client, coin_name.to_string()).await;
-
-        assert!(result_indexer.is_some());
-        let result_indexer = result_indexer.unwrap();
-        assert_eq!(result_indexer.decimals, 2);
-        assert_eq!(result_indexer.name, "Trusted Coin");
-        assert_eq!(result_indexer.symbol, "TRUSTED");
-        assert_eq!(result_indexer.description, "Trusted Coin for test");
-        assert_eq!(result_indexer.icon_url, None);
-        assert!(result_indexer.id.is_some());
-
-        let (_, result_indexer) = get_coin_metadata_fullnode_indexer(
-            cluster,
-            client,
-            immutable_metadata_coin_name.to_string(),
-        )
-        .await;
-
-        assert!(result_indexer.is_some());
-        let result_indexer = result_indexer.unwrap();
-        assert_eq!(result_indexer.decimals, 2);
-        assert_eq!(result_indexer.name, "Immutable Metadata Trusted Coin");
-        assert_eq!(result_indexer.symbol, "IMM_META_TRUSTED");
-        assert_eq!(
-            result_indexer.description,
-            "Immutable Metadata Trusted Coin for test"
-        );
-        assert_eq!(result_indexer.icon_url, None);
-        assert!(result_indexer.id.is_none()); // Immutable data is stored in struct that doesn't have ID
-    });
-}
-
-#[test]
-fn get_coin_metadata_with_native_coin_manager_coins() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-    runtime.block_on(async move {
-        let (cluster, store, client) = &start_test_cluster_with_read_write_indexer(
-            Some("get_coin_metadata_with_native_coin_manager_coins"),
-            None,
-            None,
-        )
-        .await;
-
-        let address = cluster.wallet.active_address().unwrap();
-        let address_kp = cluster
-            .wallet
-            .config()
-            .keystore()
-            .get_key(&address)
-            .unwrap();
-        let (coin_name, immutable_metadata_coin_name) =
-            create_native_coin_manager_coins(cluster, client, store, address, address_kp)
-                .await
-                .unwrap();
-
-        let (result_fullnode, result_indexer) =
-            get_coin_metadata_fullnode_indexer(cluster, client, coin_name.to_string()).await;
-
-        assert!(result_indexer.is_some());
-        assert_eq!(result_fullnode, result_indexer);
-        assert!(result_indexer.unwrap().id.is_some());
-
-        let (result_fullnode, result_indexer) = get_coin_metadata_fullnode_indexer(
-            cluster,
-            client,
-            immutable_metadata_coin_name.to_string(),
-        )
-        .await;
-
-        assert!(result_indexer.is_some());
-        assert_eq!(result_fullnode, result_indexer);
-        assert!(result_indexer.unwrap().id.is_none()); // Immutable data is stored in struct that doesn't have ID
     });
 }
 
@@ -1039,26 +939,6 @@ async fn publish_test_move_package(
         .unwrap();
 
     Ok((*package_id, tx_response))
-}
-
-async fn create_native_coin_manager_coins(
-    cluster: &TestCluster,
-    indexer_client: &HttpClient,
-    pg_store: &PgIndexerStore,
-    address: IotaAddress,
-    account_keypair: &IotaKeyPair,
-) -> Result<(String, String), anyhow::Error> {
-    let http_client = cluster.rpc_client();
-
-    let (package_id, tx_response) =
-        publish_test_move_package(http_client, address, account_keypair, "coin_manager_coins")
-            .await?;
-    indexer_wait_for_transaction(tx_response.digest, pg_store, indexer_client).await;
-
-    let coin_name = format!("{package_id}::normal_coin::NORMAL_COIN");
-    let immutable_metadata_coin_name =
-        format!("{package_id}::immutable_metadata_coin::IMMUTABLE_METADATA_COIN");
-    Ok((coin_name, immutable_metadata_coin_name))
 }
 
 async fn get_single_owned_object_by_type(
