+++ crates/iota-indexer/tests/rpc-tests/indexer_api.rs
@@ -31,9 +31,8 @@ use move_core_types::{
 };
 
 use crate::common::{
-    ApiTestSetup, indexer_wait_for_checkpoint, indexer_wait_for_latest_checkpoint,
-    indexer_wait_for_object, indexer_wait_for_transaction, rpc_call_error_msg_matches,
-    start_test_cluster_with_read_write_indexer,
+    ApiTestSetup, indexer_wait_for_checkpoint, indexer_wait_for_object,
+    indexer_wait_for_transaction, rpc_call_error_msg_matches,
 };
 
 #[test]
@@ -185,69 +184,6 @@ fn query_events_supported_events() {
             assert!(result.is_ok());
         }
     });
-}
-
-#[tokio::test]
-async fn query_validator_epoch_info_event() {
-    let (cluster, store, client) = &start_test_cluster_with_read_write_indexer(
-        Some("query_validator_epoch_info_event"),
-        None,
-        None,
-    )
-    .await;
-    indexer_wait_for_checkpoint(store, 1).await;
-
-    cluster.force_new_epoch().await;
-    indexer_wait_for_latest_checkpoint(store, cluster).await;
-
-    let result = client.query_events(EventFilter::MoveEventType("0x0000000000000000000000000000000000000000000000000000000000000003::validator_set::ValidatorEpochInfoEventV1".parse().unwrap()), None, None, None).await;
-    assert!(result.is_ok());
-    assert!(!result.unwrap().data.is_empty());
-
-    let result = client
-        .query_events(
-            EventFilter::MoveEventType(
-                "0x3::validator_set::ValidatorEpochInfoEventV1"
-                    .parse()
-                    .unwrap(),
-            ),
-            None,
-            None,
-            None,
-        )
-        .await;
-    assert!(result.is_ok());
-    assert!(!result.unwrap().data.is_empty());
-
-    let result = client
-        .query_events(
-            EventFilter::MoveEventType(
-                "0x0003::validator_set::ValidatorEpochInfoEventV1"
-                    .parse()
-                    .unwrap(),
-            ),
-            None,
-            None,
-            None,
-        )
-        .await;
-    assert!(result.is_ok());
-    assert!(!result.unwrap().data.is_empty());
-
-    let result = client
-        .query_events(
-            EventFilter::MoveEventType(
-                "0x1::validator_set::ValidatorEpochInfoEventV1"
-                    .parse()
-                    .unwrap(),
-            ),
-            None,
-            None,
-            None,
-        )
-        .await;
-    assert!(result.is_ok());
-    assert!(result.unwrap().data.is_empty());
 }
 
 #[test]
