+++ crates/iota-indexer/tests/rpc-tests/transaction_builder.rs
@@ -3,40 +3,27 @@
 
 use std::str::FromStr;
 
-use iota_indexer::store::PgIndexerStore;
 use iota_json::{call_args, type_args};
 use iota_json_rpc_api::{
-    CoinReadApiClient, GovernanceReadApiClient, IndexerApiClient, ReadApiClient,
-    TransactionBuilderClient,
+    CoinReadApiClient, GovernanceReadApiClient, ReadApiClient, TransactionBuilderClient,
 };
 use iota_json_rpc_types::{
-    IotaObjectDataOptions, IotaObjectResponseQuery, MoveCallParams, ObjectsPage,
-    RPCTransactionRequestParams, StakeStatus, TransactionBlockBytes, TransferObjectParams,
+    IotaObjectDataOptions, MoveCallParams, RPCTransactionRequestParams, StakeStatus,
+    TransactionBlockBytes, TransferObjectParams,
 };
-use iota_protocol_config::ProtocolConfig;
-use iota_swarm_config::genesis_config::AccountConfig;
 use iota_types::{
     IOTA_FRAMEWORK_ADDRESS,
-    base_types::{IotaAddress, MoveObjectType, ObjectID},
+    base_types::{IotaAddress, ObjectID},
     crypto::{AccountKeyPair, get_key_pair},
-    digests::TransactionDigest,
     gas_coin::GAS,
-    id::UID,
     iota_system_state::iota_system_state_summary::IotaSystemStateSummary,
-    object::{Data, MoveObject, OBJECT_START_VERSION, ObjectInner, Owner},
-    timelock::{
-        label::label_struct_tag_to_string, stardust_upgrade_label::stardust_upgrade_label_type,
-        timelock::TimeLock,
-    },
+    object::Owner,
 };
 use jsonrpsee::http_client::HttpClient;
 use test_cluster::TestCluster;
 
-use crate::common::{
-    ApiTestSetup, execute_tx_and_wait_for_indexer, indexer_wait_for_checkpoint,
-    indexer_wait_for_latest_checkpoint, indexer_wait_for_object,
-    start_test_cluster_with_read_write_indexer,
-};
+use crate::common::{ApiTestSetup, execute_tx_and_wait_for_indexer, indexer_wait_for_object};
+
 const FUNDED_BALANCE_PER_COIN: u64 = 10_000_000_000;
 
 #[test]
@@ -445,62 +432,6 @@ fn batch_transaction() {
 }
 
 #[test]
-fn request_add_stake() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-
-    runtime
-        .block_on(async move {
-            let (cluster, store, client) = &start_test_cluster_with_read_write_indexer(
-                Some("transaction_builder_request_add_stake"),
-                None,
-                None,
-            )
-            .await;
-            let (address, keypair): (_, AccountKeyPair) = get_key_pair();
-            let coins = create_coins_and_wait_for_indexer(cluster, client, address, 4).await;
-            let gas = coins[3];
-            let coins_to_stake = coins[..3].to_vec();
-            let validator = get_validator(client).await;
-            // subtracting some amount to see if it is possible to stake smaller amount than
-            // is provided in the input coins
-            let stake_amount = FUNDED_BALANCE_PER_COIN * 3 - 10_000;
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_add_stake(
-                    address,
-                    coins_to_stake,
-                    Some(stake_amount.into()),
-                    validator,
-                    Some(gas),
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(client, cluster, store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_stakes(address).await.unwrap();
-
-            assert_eq!(1, staked_iota.len());
-            let staked_iota = &staked_iota[0];
-            assert_eq!(validator, staked_iota.validator_address);
-
-            assert_eq!(1, staked_iota.stakes.len());
-            let stake = &staked_iota.stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Pending));
-            assert_eq!(stake.principal, stake_amount);
-
-            cluster.force_new_epoch().await;
-            indexer_wait_for_latest_checkpoint(store, cluster).await;
-            let staked_iota = client.get_stakes(address).await.unwrap();
-            let stake = &staked_iota[0].stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Active { .. }));
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .unwrap();
-}
-
-#[test]
 fn request_withdraw_stake_from_pending() {
     let ApiTestSetup {
         runtime,
@@ -556,221 +487,6 @@ fn request_withdraw_stake_from_pending() {
         .unwrap();
 }
 
-#[test]
-fn request_withdraw_stake_from_active() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-
-    runtime
-        .block_on(async move {
-            let (cluster, store, client) = &start_test_cluster_with_read_write_indexer(
-                Some("transaction_builder_request_withdraw_stake_from_active"),
-                None,
-                None,
-            )
-            .await;
-            let (address, keypair): (_, AccountKeyPair) = get_key_pair();
-            let coins = create_coins_and_wait_for_indexer(cluster, client, address, 4).await;
-            let gas = coins[3];
-            let coins_to_stake = coins[..3].to_vec();
-            let validator = get_validator(client).await;
-            // subtracting some amount to see if it is possible to stake smaller amount than
-            // is provided in the input coins
-            let stake_amount = FUNDED_BALANCE_PER_COIN * 3 - 10_000;
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_add_stake(
-                    address,
-                    coins_to_stake,
-                    Some(stake_amount.into()),
-                    validator,
-                    Some(gas),
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(client, cluster, store, tx_bytes, &keypair).await;
-
-            cluster.force_new_epoch().await;
-            indexer_wait_for_latest_checkpoint(store, cluster).await;
-            let staked_iota = client.get_stakes(address).await.unwrap();
-            let stake = &staked_iota[0].stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Active { .. }));
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_withdraw_stake(
-                    address,
-                    stake.staked_iota_id,
-                    Some(gas),
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(client, cluster, store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_stakes(address).await.unwrap();
-            assert!(staked_iota.is_empty());
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .unwrap();
-}
-
-#[test]
-fn request_add_timelocked_stake() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-
-    runtime
-        .block_on(async move {
-            let (address, keypair): (_, AccountKeyPair) = get_key_pair();
-            let (cluster, store, client, timelocked_balance) = create_cluster_with_timelocked_iota(
-                address,
-                "transaction_builder_request_add_timelocked_stake",
-            )
-            .await;
-            indexer_wait_for_checkpoint(&store, 1).await;
-
-            let coin = get_gas_object_id(&client, address).await;
-            let validator = get_validator(&client).await;
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_add_timelocked_stake(
-                    address,
-                    timelocked_balance,
-                    validator,
-                    coin,
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(&client, &cluster, &store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-
-            assert_eq!(1, staked_iota.len());
-            let staked_iota = &staked_iota[0];
-            assert_eq!(validator, staked_iota.validator_address);
-
-            assert_eq!(1, staked_iota.stakes.len());
-            let stake = &staked_iota.stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Pending));
-
-            cluster.force_new_epoch().await;
-            indexer_wait_for_latest_checkpoint(&store, &cluster).await;
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-            let stake = &staked_iota[0].stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Active { .. }));
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .unwrap();
-}
-
-#[test]
-fn request_withdraw_timelocked_stake_from_pending() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-
-    runtime
-        .block_on(async move {
-            let (address, keypair): (_, AccountKeyPair) = get_key_pair();
-            let (cluster, store, client, timelocked_balance) = create_cluster_with_timelocked_iota(
-                address,
-                "transaction_builder_request_withdraw_timelocked_stake_from_pending",
-            )
-            .await;
-            indexer_wait_for_checkpoint(&store, 1).await;
-
-            let coin = get_gas_object_id(&client, address).await;
-            let validator = get_validator(&client).await;
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_add_timelocked_stake(
-                    address,
-                    timelocked_balance,
-                    validator,
-                    coin,
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(&client, &cluster, &store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-            let stake = &staked_iota[0].stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Pending));
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_withdraw_timelocked_stake(
-                    address,
-                    stake.timelocked_staked_iota_id,
-                    coin,
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(&client, &cluster, &store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-            assert!(staked_iota.is_empty());
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .unwrap();
-}
-
-#[test]
-fn request_withdraw_timelocked_stake_from_active() {
-    let ApiTestSetup { runtime, .. } = ApiTestSetup::get_or_init();
-
-    runtime
-        .block_on(async move {
-            let (address, keypair): (_, AccountKeyPair) = get_key_pair();
-            let (cluster, store, client, timelocked_balance) = create_cluster_with_timelocked_iota(
-                address,
-                "transaction_builder_request_withdraw_timelocked_stake_from_active",
-            )
-            .await;
-            indexer_wait_for_checkpoint(&store, 1).await;
-
-            let coin = get_gas_object_id(&client, address).await;
-            let validator = get_validator(&client).await;
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_add_timelocked_stake(
-                    address,
-                    timelocked_balance,
-                    validator,
-                    coin,
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(&client, &cluster, &store, tx_bytes, &keypair).await;
-
-            cluster.force_new_epoch().await;
-            indexer_wait_for_latest_checkpoint(&store, &cluster).await;
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-            let stake = &staked_iota[0].stakes[0];
-            assert!(matches!(stake.status, StakeStatus::Active { .. }));
-
-            let tx_bytes: TransactionBlockBytes = client
-                .request_withdraw_timelocked_stake(
-                    address,
-                    stake.timelocked_staked_iota_id,
-                    coin,
-                    100_000_000.into(),
-                )
-                .await
-                .unwrap();
-            execute_tx_and_wait_for_indexer(&client, &cluster, &store, tx_bytes, &keypair).await;
-
-            let staked_iota = client.get_timelocked_stakes(address).await.unwrap();
-            assert!(staked_iota.is_empty());
-
-            Ok::<(), anyhow::Error>(())
-        })
-        .unwrap();
-}
-
 async fn get_address_balances(indexer_client: &HttpClient, address: IotaAddress) -> Vec<u64> {
     indexer_client
         .get_coins(address, None, None, None)
@@ -803,80 +519,6 @@ async fn create_coins_and_wait_for_indexer(
     coins
 }
 
-async fn create_cluster_with_timelocked_iota(
-    address: IotaAddress,
-    indexer_db_name: &str,
-) -> (TestCluster, PgIndexerStore, HttpClient, ObjectID) {
-    let principal = 100_000_000_000;
-    let expiration_timestamp_ms = u64::MAX;
-    let label = Option::Some(label_struct_tag_to_string(stardust_upgrade_label_type()));
-
-    let timelock_iota = {
-        MoveObject::new_from_execution(
-            MoveObjectType::timelocked_iota_balance(),
-            OBJECT_START_VERSION,
-            TimeLock::<iota_types::balance::Balance>::new(
-                UID::new(ObjectID::random()),
-                iota_types::balance::Balance::new(principal),
-                expiration_timestamp_ms,
-                label.clone(),
-            )
-            .to_bcs_bytes(),
-            &ProtocolConfig::get_for_min_version(),
-        )
-        .unwrap()
-    };
-    let timelock_iota = ObjectInner {
-        owner: Owner::AddressOwner(address),
-        data: Data::Move(timelock_iota),
-        previous_transaction: TransactionDigest::genesis_marker(),
-        storage_rebate: 0,
-    };
-
-    let (cluster, store, client) = start_test_cluster_with_read_write_indexer(
-        Some(indexer_db_name),
-        Some(Box::new(move |builder| {
-            builder
-                .with_accounts(
-                    [AccountConfig {
-                        address: Some(address),
-                        gas_amounts: [1_000_000_000].into(),
-                    }]
-                    .into(),
-                )
-                .with_objects([timelock_iota.into()])
-        })),
-        None,
-    )
-    .await;
-
-    let fullnode_client = cluster.rpc_client();
-
-    let objects: ObjectsPage = fullnode_client
-        .get_owned_objects(
-            address,
-            Some(IotaObjectResponseQuery::new_with_options(
-                IotaObjectDataOptions::full_content(),
-            )),
-            None,
-            None,
-        )
-        .await
-        .unwrap();
-    assert_eq!(2, objects.data.len());
-
-    let timelocked_balance = objects
-        .data
-        .into_iter()
-        .find(|o| !o.data.as_ref().unwrap().is_gas_coin())
-        .unwrap()
-        .object()
-        .unwrap()
-        .object_id;
-
-    (cluster, store, client, timelocked_balance)
-}
-
 async fn get_validator(client: &HttpClient) -> IotaAddress {
     let iota_system_state = client.get_latest_iota_system_state_v2().await.unwrap();
     match iota_system_state {
@@ -884,13 +526,4 @@ async fn get_validator(client: &HttpClient) -> IotaAddress {
         IotaSystemStateSummary::V2(v2) => v2.active_validators[0].iota_address,
         _ => panic!("unsupported IotaSystemStateSummary"),
     }
-}
-
-async fn get_gas_object_id(client: &HttpClient, address: IotaAddress) -> ObjectID {
-    client
-        .get_coins(address, None, None, None)
-        .await
-        .unwrap()
-        .data[0]
-        .coin_object_id
 }
